+++ packages/xchain-ethereum/__tests__/client.test.ts
@@ -1,11 +1,11 @@
-import Client from '../src/client'
 import { ethers, Wallet, providers } from 'ethers'
 import { formatEther, parseEther } from '@ethersproject/units'
 import { TransactionResponse, TransactionReceipt } from '@ethersproject/abstract-provider'
-import * as blockChair from '../src/blockchair-api'
-import { AddressDTO } from '@xchainjs/xchain-ethereum/src/types/blockchair-api-types'
 import { baseAmount, AssetETH, BaseAmount } from '@xchainjs/xchain-util'
 import { Balances } from '@xchainjs/xchain-client'
+import Client from '../src/client'
+import { ETH_DECIMAL } from '../src/utils'
+import { mockDashboardAddress, mockGetTx } from '../__mocks__/blockchair-api'
 
 /**
  * Test Data
@@ -13,8 +13,9 @@ import { Balances } from '@xchainjs/xchain-client'
  */
 const phrase = 'canyon throw labor waste awful century ugly they found post source draft'
 const newPhrase = 'logic neutral rug brain pluck submit earth exit erode august remain ready'
-const address = '0xb8c0c226d6FE17E5d9132741836C3ae82A5B6C4E'
+const address = '0xb8c0c226d6fe17e5d9132741836c3ae82a5b6c4e'
 const vault = '0x8c2A90D36Ec9F745C9B28B588Cba5e2A978A1656'
+const blockchairUrl = 'https://api.blockchair.com/ethereum/testnet'
 const wallet = {
   signingKey: {
     curve: 'secp256k1',
@@ -35,8 +36,8 @@ const txResponse = {
   difficulty: 2,
   gasLimit: ethers.BigNumber.from('0x47e7c4'),
   gasPrice: ethers.BigNumber.from('0x3b9aca00'),
-  to: '0xb8c0c226d6FE17E5d9132741836C3ae82A5B6C4E',
-  from: '0xb8c0c226d6FE17E5d9132741836C3ae82A5B6C4E',
+  to: '0xb8c0c226d6fe17e5d9132741836c3ae82a5b6c4e',
+  from: '0xb8c0c226d6fe17e5d9132741836c3ae82a5b6c4e',
   value: ethers.BigNumber.from('0x00'),
   data: '0x54455354',
   chainId: 4,
@@ -141,17 +142,273 @@ describe('Utils', () => {
   })
 })
 
-describe('Transactions', () => {
-  it('gets transaction count', async () => {
+describe('Balances', () => {
+  it('gets a balance without address args', async () => {
+    const ethClient = new Client({
+      network: 'testnet',
+      phrase,
+      blockchairUrl,
+    })
+
+    mockDashboardAddress(blockchairUrl, ethClient.getAddress(), {
+      '0xb8c0c226d6fe17e5d9132741836c3ae82a5b6c4e': {
+        address: {
+          type: 'account',
+          contract_code_hex: null,
+          contract_created: null,
+          contract_destroyed: null,
+          balance: '100',
+          balance_usd: 0,
+          received_approximate: '4358301000000000000',
+          received_usd: 0,
+          spent_approximate: '0',
+          spent_usd: 0,
+          fees_approximate: '0',
+          fees_usd: 0,
+          receiving_call_count: 38,
+          spending_call_count: 0,
+          call_count: 38,
+          transaction_count: 38,
+          first_seen_receiving: '2020-04-21 07:52:18',
+          last_seen_receiving: '2020-12-08 04:49:56',
+          first_seen_spending: null,
+          last_seen_spending: null,
+          nonce: null
+        },
+        calls: [],
+      }
+    })
+
+    const balance = await ethClient.getBalance()
+    expect(balance.length).toEqual(1)
+    expect(balance[0].asset).toEqual(AssetETH)
+    expect(balance[0].amount.amount().isEqualTo(baseAmount(100, ETH_DECIMAL).amount())).toBeTruthy()
+  })
+
+  it('gets a balance from address', async () => {
+    const ethClient = new Client({
+      network: 'testnet',
+      phrase,
+      blockchairUrl,
+    })
+
+    mockDashboardAddress(blockchairUrl, '0xdac17f958d2ee523a2206206994597c13d831ec7', {
+      '0xdac17f958d2ee523a2206206994597c13d831ec7': {
+        address: {
+          type: 'account',
+          contract_code_hex: null,
+          contract_created: null,
+          contract_destroyed: null,
+          balance: '10000',
+          balance_usd: 0,
+          received_approximate: '4358301000000000000',
+          received_usd: 0,
+          spent_approximate: '0',
+          spent_usd: 0,
+          fees_approximate: '0',
+          fees_usd: 0,
+          receiving_call_count: 38,
+          spending_call_count: 0,
+          call_count: 38,
+          transaction_count: 38,
+          first_seen_receiving: '2020-04-21 07:52:18',
+          last_seen_receiving: '2020-12-08 04:49:56',
+          first_seen_spending: null,
+          last_seen_spending: null,
+          nonce: null
+        },
+        calls: [],
+      }
+    })
+
+    const balance = await ethClient.getBalance('0xdac17f958d2ee523a2206206994597c13d831ec7')
+    expect(balance.length).toEqual(1)
+    expect(balance[0].asset).toEqual(AssetETH)
+    expect(balance[0].amount.amount().isEqualTo(baseAmount(10000, ETH_DECIMAL).amount())).toBeTruthy()
+  })
+
+  it('throws error on bad address', async () => {
     const ethClient = new Client({ network: 'testnet', phrase })
 
-    const mockTxCount = jest.spyOn(ethClient.getProvider(), 'getTransactionCount')
-    mockTxCount.mockImplementation(async (_): Promise<number> => Promise.resolve(1))
+    const balances = ethClient.getBalance('0xbad')
+    expect(balances).rejects.toThrowError()
+  })
+})
 
-    const count = await ethClient.getTransactionCount()
+describe('Transactions', () => {
+  it('get transaction history', async () => {
+    const ethClient = new Client({
+      network: 'testnet',
+      phrase,
+      blockchairUrl,
+    })
 
-    expect(mockTxCount).toHaveBeenCalledWith(ethClient.getAddress(), 'latest')
-    expect(count).toEqual(1)
+    mockDashboardAddress(blockchairUrl, '0xdac17f958d2ee523a2206206994597c13d831ec7', {
+      '0xdac17f958d2ee523a2206206994597c13d831ec7': {
+        address: {
+          type: 'account',
+          contract_code_hex: null,
+          contract_created: null,
+          contract_destroyed: null,
+          balance: '10000',
+          balance_usd: 0,
+          received_approximate: '4358301000000000000',
+          received_usd: 0,
+          spent_approximate: '0',
+          spent_usd: 0,
+          fees_approximate: '0',
+          fees_usd: 0,
+          receiving_call_count: 38,
+          spending_call_count: 0,
+          call_count: 38,
+          transaction_count: 38,
+          first_seen_receiving: '2020-04-21 07:52:18',
+          last_seen_receiving: '2020-12-08 04:49:56',
+          first_seen_spending: null,
+          last_seen_spending: null,
+          nonce: null,
+        },
+        calls: [{
+          block_id: 3888697,
+          transaction_hash: '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2',
+          index: '0',
+          time: '2020-12-08 04:49:56',
+          sender: '0x4a89644d5dffb825a42a3496e24510424ca01516',
+          recipient: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+          value: 0,
+          value_usd: null,
+          transferred: true,
+        },{
+          block_id: 3888697,
+          transaction_hash: '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2',
+          index: '0',
+          time: '2020-12-08 04:49:56',
+          sender: '0x4a89644d5dffb825a42a3496e24510424ca01516',
+          recipient: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+          value: 0,
+          value_usd: null,
+          transferred: true,
+        }],
+      }
+    })
+
+    mockDashboardAddress(blockchairUrl, '0xdac17f958d2ee523a2206206994597c13d831ec7', {
+      '0xdac17f958d2ee523a2206206994597c13d831ec7': {
+        address: {
+          type: 'account',
+          contract_code_hex: null,
+          contract_created: null,
+          contract_destroyed: null,
+          balance: '10000',
+          balance_usd: 0,
+          received_approximate: '4358301000000000000',
+          received_usd: 0,
+          spent_approximate: '0',
+          spent_usd: 0,
+          fees_approximate: '0',
+          fees_usd: 0,
+          receiving_call_count: 38,
+          spending_call_count: 0,
+          call_count: 38,
+          transaction_count: 38,
+          first_seen_receiving: '2020-04-21 07:52:18',
+          last_seen_receiving: '2020-12-08 04:49:56',
+          first_seen_spending: null,
+          last_seen_spending: null,
+          nonce: null,
+        },
+        calls: [{
+          block_id: 3888697,
+          transaction_hash: '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2',
+          index: '0',
+          time: '2020-12-08 04:49:56',
+          sender: '0x4a89644d5dffb825a42a3496e24510424ca01516',
+          recipient: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+          value: 1000,
+          value_usd: null,
+          transferred: true,
+        }],
+      }
+    })
+
+    const txHistory = await ethClient.getTransactions({address: '0xdac17f958d2ee523a2206206994597c13d831ec7', limit: 1})
+    expect(txHistory.total).toEqual(2)
+    expect(txHistory.txs[0].hash).toEqual('0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2')
+    expect(txHistory.txs[0].asset).toEqual(AssetETH)
+    expect(txHistory.txs[0].from[0].from).toEqual('0x4a89644d5dffb825a42a3496e24510424ca01516')
+    expect(txHistory.txs[0].from[0].amount.amount().isEqualTo(baseAmount(1000, ETH_DECIMAL).amount())).toBeTruthy()
+    expect(txHistory.txs[0].to[0].to).toEqual('0xdac17f958d2ee523a2206206994597c13d831ec7')
+    expect(txHistory.txs[0].to[0].amount.amount().isEqualTo(baseAmount(1000, ETH_DECIMAL).amount())).toBeTruthy()
+    expect(txHistory.txs[0].type).toEqual('transfer')
+  })
+
+  it('get transaction data', async () => {
+    const ethClient = new Client({
+      network: 'testnet',
+      phrase,
+      blockchairUrl,
+    })
+
+    mockGetTx(blockchairUrl, '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2', {
+      '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2': {
+        transaction: {
+          block_id: 3888697,
+          id: 3888697000001,
+          index: 1,
+          hash: '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2',
+          date: '2020-12-08',
+          time: '2020-12-08 04:49:56',
+          failed: false,
+          type: 'call',
+          sender: '0x4a89644d5dffb825a42a3496e24510424ca01516',
+          recipient: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+          call_count: 1,
+          value: '1000',
+          value_usd: 0,
+          internal_value: '0',
+          internal_value_usd: null,
+          fee: '432160000000000',
+          fee_usd: null,
+          gas_used: 21608,
+          gas_limit: 32412,
+          gas_price: 20000000000,
+          input_hex: 'a9059cbb0000000000000000000000001d5516bc994a130fc4fa2a6275f5b4ee022b18720000000000000000000000000000000000000000000000000000000004cace80',
+          nonce: 98,
+          v: '2e',
+          r: '8797fe82968c48ccaf7db43785a2a97c6868b187a009f930e8406a052aefdc7b',
+          s: '54dc9740ebf75835bd3aca3de018454251213e13b44fa93ecbf5f17faf16d995'
+        },
+        calls: [{
+          block_id: 3888697,
+          transaction_id: 3888697000001,
+          transaction_hash: '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2',
+          index: '0',
+          depth: 0,
+          date: '2020-12-08',
+          time: '2020-12-08 04:49:56',
+          failed: false,
+          fail_reason: null,
+          type: 'call',
+          sender: '0x4a89644d5dffb825a42a3496e24510424ca01516',
+          recipient: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+          child_call_count: 0,
+          value: '1000',
+          value_usd: 0,
+          transferred: true,
+          input_hex: 'a9059cbb0000000000000000000000001d5516bc994a130fc4fa2a6275f5b4ee022b18720000000000000000000000000000000000000000000000000000000004cace80',
+          output_hex: ''
+        }],
+      }
+    })
+
+    const txData = await ethClient.getTransactionData('0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2')
+    expect(txData.hash).toEqual('0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2')
+    expect(txData.asset).toEqual(AssetETH)
+    expect(txData.from[0].from).toEqual('0x4a89644d5dffb825a42a3496e24510424ca01516')
+    expect(txData.from[0].amount.amount().isEqualTo(baseAmount(1000, ETH_DECIMAL).amount())).toBeTruthy()
+    expect(txData.to[0].to).toEqual('0xdac17f958d2ee523a2206206994597c13d831ec7')
+    expect(txData.to[0].amount.amount().isEqualTo(baseAmount(1000, ETH_DECIMAL).amount())).toBeTruthy()
+    expect(txData.type).toEqual('transfer')
   })
 
   it('checks vault and vaultTx', async () => {
@@ -175,7 +432,7 @@ describe('Transactions', () => {
 
     await ethClient.normalTx({ recipient: ethClient.getAddress(), amount: baseAmount(parseEther('1').toString(), 18) })
     expect(mockTx).toHaveBeenCalledWith({
-      to: '0xb8c0c226d6FE17E5d9132741836C3ae82A5B6C4E',
+      to: '0xb8c0c226d6fe17e5d9132741836c3ae82a5b6c4e',
       value: '1',
     })
   })
@@ -201,67 +458,12 @@ describe('Transactions', () => {
     expect(mockTx).toHaveBeenCalledWith({
       data: '0xdeadbeef',
       nonce: 123,
-      to: '0xb8c0c226d6FE17E5d9132741836C3ae82A5B6C4E',
+      to: '0xb8c0c226d6fe17e5d9132741836c3ae82a5b6c4e',
       value: '1',
     })
   })
 })
 
-describe('Balances', () => {
-  it('gets a balance without address args', async () => {
-    const ethClient = new Client({
-      network: 'testnet',
-      phrase,
-      blockchairUrl: 'urlHere',
-      blockchairNodeApiKey: 'etherKey',
-    })
-
-    const mockBalance = jest.spyOn(blockChair, 'getAddress')
-    mockBalance.mockImplementation(async () => {
-      return Promise.resolve(({
-        [address]: {
-          address: {
-            balance: parseEther('1.786'),
-          },
-        },
-      } as unknown) as AddressDTO)
-    })
-
-    await ethClient.getBalance()
-
-    expect(mockBalance).toHaveBeenCalledWith('urlHere', address, 'etherKey')
-  })
-
-  it('gets a balance from address', async () => {
-    const ethClient = new Client({
-      network: 'testnet',
-      phrase,
-      blockchairUrl: 'urlHere',
-      blockchairNodeApiKey: 'etherKey',
-    })
-
-    const mockBalance = jest.spyOn(blockChair, 'getAddress')
-    mockBalance.mockImplementation(async () => {
-      return Promise.resolve(({
-        ['0xb1d133e115E32Bee0F163EcD2c60FB462b8cDdC1']: {
-          address: {
-            balance: parseEther('1.786'),
-          },
-        },
-      } as unknown) as AddressDTO)
-    })
-
-    await ethClient.getBalance('0xb1d133e115E32Bee0F163EcD2c60FB462b8cDdC1')
-
-    expect(mockBalance).toHaveBeenCalledWith('urlHere', '0xb1d133e115E32Bee0F163EcD2c60FB462b8cDdC1', 'etherKey')
-  })
-
-  it('throws error on bad address', async () => {
-    const ethClient = new Client({ network: 'testnet', phrase })
-    const balances = ethClient.getBalance('0xbad')
-    expect(balances).rejects.toThrowError()
-  })
-})
 
 describe('ERC20', () => {
   it('gets erc 20 balance for a contract without addr', async () => {